Move new ITs out of test
